+++ src/onegov.core/HISTORY.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-- Adds Morepath 0.13 compatibility.
+- Fixes meta/content failing if the dictionary is None.
   [href]
 
 0.18.1 (2016-04-01)