+++ src/onegov.core/HISTORY.rst
@@ -1,5 +1,7 @@
 Changelog
 ---------
+0.18.2 (2016-04-05)
+~~~~~~~~~~~~~~~~~~~
 
 - Fixes meta/content failing if the dictionary is None.
   [href]